Arguments in 'Text' were converted to double. Fix for MATLAB 2014b compatibility.
git-svn-id: https://sccn.ucsd.edu/svn/software/eeglab@10441 1e58238c-3a72-4c0f-aabe-36d030380f39

+++ functions/sigprocfunc/erpimage.m
@@ -3093,10 +3093,10 @@ if ~isnan(coherfreq)
             ynumoffset  = double(limit(1)-0.3*ydelta); % double for Matlab 7
         end
         
-        t=text(ynumoffset,maxampERP,num2str(maxampERP,3));
+        t=text(double(ynumoffset),double(maxampERP),num2str(maxampERP,3));
         set(t,'HorizontalAlignment','right','FontSize',TICKFONT);
         
-        t=text(ynumoffset,minampERP, num2str(minampERP,3));
+        t=text(double(ynumoffset),double(minampERP), num2str(minampERP,3));
         set(t,'HorizontalAlignment','right','FontSize',TICKFONT);
         
         ax3=axes('Position',...
